Fix an argspec detection edge case in `pecan.util.getargspec`
In lieu of decorators that responsibly duplicate the argspec of underlying
functions, pecan must make a best guess.  When multiple nested decorators are
discovered, the __closure__ of the wrapped function is searched for callables
(one of which likely represents the underlying wrapped function - the
controller method itself).  This change tweaks the algorithm a bit further to
place higher match precedence on in-scope functions that have `self` as a first
argument.