JSON-formatted HTTP response bodies should use `pecan.jsonify.encode`.
If a webob.HTTPException is raised, and the Accept header is application/json,
pecan automatically returns a JSON-formatted response body.  If this is the
case, `pecan.jsonify.encode` should be used (instead of `json.dumps`) so that
JSON serialization rules (http://pecan.readthedocs.io/en/latest/jsonify.html)
are respected.