routing: stop printing warning if the user passes a wrong URL
When a _lookup method accepts N + *remainder arguments but the number of
path components (separated by /) is < N, a TypeError is raised when calling the
object and Pecan prints a RuntimeWarning that is log. Unfortunately this just
fills up the log with no interesting information for the developer as this is
just a standard 404 error.